pyproject.toml
[tool.poetry]
name = "django-inertia-vite"
version = "0.1.0"
description = ""
authors = ["jasonxie <x5758x@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.1.4"
django-vite = "^2.0.2"
inertia-django = "^0.4.1"
whitenoise = "^6.2.0"

[tool.poetry.group.dev.dependencies]
black = "^22.10.0"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
